+++ MAINTAINERS.yml
@@ -2582,7 +2582,7 @@ Documentation Infrastructure:
   files:
     - drivers/video/
     - include/zephyr/drivers/video.h
-    - include/zephyr/drivers/video-controls.h
+    - include/zephyr/drivers/video/
     - include/zephyr/dt-bindings/video/
     - doc/hardware/peripherals/video.rst
     - tests/drivers/*/video/
@@ -5721,6 +5721,21 @@ VFS:
     - "area: File System"
   tests:
     - filesystem
+
+Video Subsystem:
+  status: maintained
+  maintainers:
+    - josuah
+    - ngphibang
+  collaborators:
+    - loicpoulain
+    - avolmat-st
+  files:
+    - include/zephyr/video/
+  labels:
+    - "area: Video"
+  tests:
+    - video
 
 WCH Platforms:
   status: maintained
