+++ drivers/video/mt9m114.c
@@ -7,13 +7,14 @@
 
 #define DT_DRV_COMPAT aptina_mt9m114
 
-#include <zephyr/kernel.h>
 #include <zephyr/device.h>
+#include <zephyr/drivers/i2c.h>
+#include <zephyr/drivers/video.h>
+#include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/byteorder.h>
-#include <zephyr/drivers/video.h>
-#include <zephyr/drivers/video-controls.h>
-#include <zephyr/drivers/i2c.h>
+#include <zephyr/video/controls.h>
+#include <zephyr/video/formats.h>
 
 #include "video_ctrls.h"
 #include "video_device.h"
