+++ drivers/video/video_stm32_dcmipp.c
@@ -7,17 +7,17 @@
 #include <errno.h>
 
 #include <stm32_bitops.h>
-#include <zephyr/kernel.h>
-#include <zephyr/irq.h>
-#include <zephyr/drivers/clock_control/stm32_clock_control.h>
 #include <zephyr/drivers/clock_control.h>
+#include <zephyr/drivers/clock_control/stm32_clock_control.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/drivers/reset.h>
 #include <zephyr/drivers/video.h>
-#include <zephyr/drivers/video-controls.h>
-#include <zephyr/dt-bindings/video/video-interfaces.h>
-#include <zephyr/logging/log.h>
 #include <zephyr/drivers/video/stm32_dcmipp.h>
+#include <zephyr/dt-bindings/video/video-interfaces.h>
+#include <zephyr/irq.h>
+#include <zephyr/kernel.h>
+#include <zephyr/logging/log.h>
+#include <zephyr/video/video.h>
 
 #include "video_ctrls.h"
 #include "video_device.h"
