+++ include/zephyr/drivers/video.h
@@ -2,7 +2,7 @@
  * Copyright (c) 2019 Linaro Limited.
  * Copyright 2025 NXP
  * Copyright (c) 2025 STMicroelectronics
- *
+ * SPDX-FileCopyrightText: Copyright The Zephyr Project Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -12,8 +12,8 @@
  * @brief Main header file for video driver API.
  */
 
-#ifndef ZEPHYR_INCLUDE_VIDEO_H_
-#define ZEPHYR_INCLUDE_VIDEO_H_
+#ifndef ZEPHYR_INCLUDE_DRIVERS_VIDEO_H
+#define ZEPHYR_INCLUDE_DRIVERS_VIDEO_H
 
 /**
  * @brief Interfaces for video devices.
@@ -24,356 +24,69 @@
  * @{
  */
 
-#include <zephyr/device.h>
 #include <stddef.h>
-#include <zephyr/kernel.h>
+#include <stdint.h>
 
+#include <zephyr/device.h>
+#include <zephyr/kernel.h>
 #include <zephyr/types.h>
+#include <zephyr/video/types.h>
+#include <zephyr/video/controls.h>
+#include <zephyr/video/formats.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-struct video_control;
-
 /**
- * @brief video_buf_type enum
- *
- * Supported video buffer types of a video device.
- * The direction (input or output) is defined from the device's point of view.
- * Devices like cameras support only output type, encoders support only input
- * types while m2m devices like ISP, PxP support both input and output types.
- */
-enum video_buf_type {
-	/** input buffer type */
-	VIDEO_BUF_TYPE_INPUT = 1,
-	/** output buffer type */
-	VIDEO_BUF_TYPE_OUTPUT = 2,
-};
-
-/**
- * @brief video_buf_memory enum
- *
- * Memory type of a buffer
- */
-enum video_buf_memory {
-	/** buffer is allocated from the internal video heap */
-	VIDEO_MEMORY_INTERNAL = 1,
-	/** buffer is allocated from outside */
-	VIDEO_MEMORY_EXTERNAL = 2,
-};
-
-/**
- * @brief Video format structure
- *
- * Used to configure frame format.
- */
-struct video_format {
-	/** type of the buffer */
-	enum video_buf_type type;
-	/** FourCC pixel format value (\ref video_pixel_formats) */
-	uint32_t pixelformat;
-	/** frame width in pixels. */
-	uint32_t width;
-	/** frame height in pixels. */
-	uint32_t height;
-	/**
-	 * @brief line stride.
-	 *
-	 * This is the number of bytes that needs to be added to the address in the
-	 * first pixel of a row in order to go to the address of the first pixel of
-	 * the next row (>=width).
-	 */
-	uint32_t pitch;
-	/**
-	 * @brief size of the buffer in bytes, need to be set by the drivers
-	 *
-	 * For uncompressed formats, this is the size of the raw data buffer in bytes,
-	 * which could be the whole raw image or a portion of the raw image in cases
-	 * the receiver / dma supports it.
-	 *
-	 * For compressed formats, this is the maximum number of bytes required to
-	 * hold a complete compressed frame, estimated for the worst case.
-	 */
-	uint32_t size;
-};
-
-/**
- * @brief Video format capability
- *
- * Used to describe a video endpoint format capability.
- */
-struct video_format_cap {
-	/** FourCC pixel format value (\ref video_pixel_formats). */
-	uint32_t pixelformat;
-	/** minimum supported frame width in pixels. */
-	uint32_t width_min;
-	/** maximum supported frame width in pixels. */
-	uint32_t width_max;
-	/** minimum supported frame height in pixels. */
-	uint32_t height_min;
-	/** maximum supported frame height in pixels. */
-	uint32_t height_max;
-	/** width step size in pixels. */
-	uint16_t width_step;
-	/** height step size in pixels. */
-	uint16_t height_step;
-};
-
-/**
- * @brief Video format capabilities
- *
- * Used to describe video endpoint capabilities.
- */
-struct video_caps {
-	/** type of the buffer */
-	enum video_buf_type type;
-	/** list of video format capabilities (zero terminated). */
-	const struct video_format_cap *format_caps;
-	/** minimal count of video buffers to enqueue before being able to start
-	 * the stream.
-	 */
-	uint8_t min_vbuf_count;
-	/** requirement on the buffer alignment, in bytes */
-	size_t buf_align;
-};
-
-/**
- * @brief Video buffer structure
- *
- * Represents a video frame.
- */
-struct video_buffer {
-	/** Pointer to driver specific data. */
-	/* It must be kept as first field of the struct if used for @ref k_fifo APIs. */
-	void *driver_data;
-	/** type of the buffer */
-	enum video_buf_type type;
-	/** type of the buffer memory, see @ref video_buf_memory */
-	uint8_t memory;
-	/** pointer to the start of the buffer. */
-	uint8_t *buffer;
-	/** index of the buffer in the video buffer pool */
-	uint16_t index;
-	/** size of the buffer in bytes. */
-	uint32_t size;
-	/** number of bytes occupied by the valid data in the buffer. */
-	uint32_t bytesused;
-	/** time reference in milliseconds at which the last data byte was
-	 * actually received for input endpoints or to be consumed for output
-	 * endpoints.
-	 */
-	uint32_t timestamp;
-	/** Line offset within frame this buffer represents, from the
-	 * beginning of the frame. This offset is given in pixels,
-	 * so `line_offset` * `pitch` provides offset from the start of
-	 * the frame in bytes.
-	 */
-	uint16_t line_offset;
-};
-
-/**
- * @brief Supported frame interval type of a video device.
- */
-enum video_frmival_type {
-	/** discrete frame interval type */
-	VIDEO_FRMIVAL_TYPE_DISCRETE = 1,
-	/** stepwise frame interval type */
-	VIDEO_FRMIVAL_TYPE_STEPWISE = 2,
-};
-
-/**
- * @brief Video frame interval structure
- *
- * Used to describe a video frame interval.
- */
-struct video_frmival {
-	/** numerator of the frame interval */
-	uint32_t numerator;
-	/** denominator of the frame interval */
-	uint32_t denominator;
-};
-
-/**
- * @brief Video frame interval stepwise structure
- *
- * Used to describe the video frame interval stepwise type.
- */
-struct video_frmival_stepwise {
-	/** minimum frame interval in seconds */
-	struct video_frmival min;
-	/** maximum frame interval in seconds */
-	struct video_frmival max;
-	/** frame interval step size in seconds */
-	struct video_frmival step;
-};
-
-/**
- * @brief Video frame interval enumeration structure
- *
- * Used to describe the supported video frame intervals of a given video format.
- */
-struct video_frmival_enum {
-	/** frame interval index during enumeration */
-	uint32_t index;
-	/** video format for which the query is made */
-	const struct video_format *format;
-	/** frame interval type the device supports */
-	enum video_frmival_type type;
-	/** the actual frame interval */
-	union {
-		struct video_frmival discrete;
-		struct video_frmival_stepwise stepwise;
-	};
-};
-
-/**
- * @brief Video signal result
- *
- * Identify video event.
- */
-enum video_signal_result {
-	VIDEO_BUF_DONE,    /**< Buffer is done */
-	VIDEO_BUF_ABORTED, /**< Buffer is aborted */
-	VIDEO_BUF_ERROR,   /**< Buffer is in error */
-};
-
-/**
- * @brief Video selection target
- *
- * Used to indicate which selection to query or set on a video device
- */
-enum video_selection_target {
-	/** Current crop setting */
-	VIDEO_SEL_TGT_CROP,
-	/** Crop bound (aka the maximum crop achievable) */
-	VIDEO_SEL_TGT_CROP_BOUND,
-	/** Native size of the input frame */
-	VIDEO_SEL_TGT_NATIVE_SIZE,
-	/** Current compose setting */
-	VIDEO_SEL_TGT_COMPOSE,
-	/** Compose bound (aka the maximum compose achievable) */
-	VIDEO_SEL_TGT_COMPOSE_BOUND,
-};
-
-/**
- * @brief Description of a rectangle area.
- *
- * Used for crop/compose and possibly within drivers as well
- */
-struct video_rect {
-	/** left offset of selection rectangle */
-	uint32_t left;
-	/** top offset of selection rectangle */
-	uint32_t top;
-	/** width of selection rectangle */
-	uint32_t width;
-	/** height of selection rectangle */
-	uint32_t height;
-};
-
-/**
- * @brief Video selection (crop / compose) structure
- *
- * Used to describe the query and set selection target on a video device
- */
-struct video_selection {
-	/** buffer type, allow to select for device having both input and output */
-	enum video_buf_type type;
-	/** selection target enum */
-	enum video_selection_target target;
-	/** selection target rectangle */
-	struct video_rect rect;
-};
-
-/**
- * @typedef video_api_format_t
- * @brief Function pointer type for video_set/get_format()
- *
- * See video_set/get_format() for argument descriptions.
+ * @brief Function type of @ref video_driver_set_format(), @ref driver_get_format()
  */
 typedef int (*video_api_format_t)(const struct device *dev, struct video_format *fmt);
 
 /**
- * @typedef video_api_frmival_t
- * @brief Function pointer type for video_set/get_frmival()
- *
- * See video_set/get_frmival() for argument descriptions.
+ * @brief Function type of @ref video_driver_set_frmival(), @ref video_driver_get_frmival()
  */
 typedef int (*video_api_frmival_t)(const struct device *dev, struct video_frmival *frmival);
 
 /**
- * @typedef video_api_enum_frmival_t
- * @brief List all supported frame intervals of a given format
- *
- * See video_enum_frmival() for argument descriptions.
+ * @brief Function type of @ref video_driver_enum_frmival()
  */
 typedef int (*video_api_enum_frmival_t)(const struct device *dev, struct video_frmival_enum *fie);
 
 /**
- * @typedef video_api_enqueue_t
- * @brief Enqueue a buffer in the driver’s incoming queue.
- *
- * See video_enqueue() for argument descriptions.
+ * @brief Function type of @ref video_driver_enqueue()
  */
 typedef int (*video_api_enqueue_t)(const struct device *dev, struct video_buffer *buf);
 
 /**
- * @typedef video_api_dequeue_t
- * @brief Dequeue a buffer from the driver’s outgoing queue.
- *
- * See video_dequeue() for argument descriptions.
+ * @brief Function type of @ref video_driver_dequeue()
  */
 typedef int (*video_api_dequeue_t)(const struct device *dev, struct video_buffer **buf,
 				   k_timeout_t timeout);
 
 /**
- * @typedef video_api_flush_t
- * @brief Flush endpoint buffers, buffer are moved from incoming queue to
- *        outgoing queue.
- *
- * See video_flush() for argument descriptions.
+ * @brief Function type of @ref video_driver_flush()
  */
 typedef int (*video_api_flush_t)(const struct device *dev, bool cancel);
 
 /**
- * @typedef video_api_set_stream_t
- * @brief Start or stop streaming on the video device.
- *
- * Start (enable == true) or stop (enable == false) streaming on the video device.
- *
- * @param dev Pointer to the device structure.
- * @param enable If true, start streaming, otherwise stop streaming.
- * @param type The type of the buffers stream to start or stop.
- *
- * @retval 0 on success, otherwise a negative errno code.
+ * @brief Function type of @ref video_driver_get_caps()
  */
 typedef int (*video_api_set_stream_t)(const struct device *dev, bool enable,
 				      enum video_buf_type type);
 
 /**
- * @typedef video_api_ctrl_t
- * @brief Set/Get a video control value.
- *
- * @param dev Pointer to the device structure.
- * @param cid Id of the control to set/get its value.
+ * @brief Function type of @ref video_driver_set_ctrl(), @ref video_driver_get_volatile_ctrl()
  */
 typedef int (*video_api_ctrl_t)(const struct device *dev, uint32_t cid);
 
 /**
- * @typedef video_api_get_caps_t
- * @brief Get capabilities of a video endpoint.
- *
- * See video_get_caps() for argument descriptions.
+ * @brief Function type of @ref video_driver_get_caps()
  */
 typedef int (*video_api_get_caps_t)(const struct device *dev, struct video_caps *caps);
 
 /**
- * @typedef video_api_transform_cap_t
- * @brief Function pointer type for transforming a @ref video_format_cap from one end
- * to the other end of a m2m video device.
- *
- * See @ref video_transform_cap for argument descriptions.
+ * @brief Function type of @ref video_driver_transform_cap()
  */
 typedef int (*video_api_transform_cap_t)(const struct device *const dev,
 					 const struct video_format_cap *const cap,
@@ -381,65 +94,70 @@ typedef int (*video_api_transform_cap_t)(const struct device *const dev,
 					 enum video_buf_type type, uint16_t ind);
 
 /**
- * @typedef video_api_set_signal_t
- * @brief Register/Unregister poll signal for buffer events.
- *
- * See video_set_signal() for argument descriptions.
+ * @brief Function type of @ref video_driver_set_signal()
  */
 typedef int (*video_api_set_signal_t)(const struct device *dev, struct k_poll_signal *sig);
 
 /**
- * @typedef video_api_selection_t
- * @brief Get/Set video selection (crop / compose)
- *
- * See @ref video_set_selection and @ref video_get_selection for argument descriptions.
+ * @brief Function type of @ref video_driver_set_selection(), @ref video_driver_get_selection()
  */
 typedef int (*video_api_selection_t)(const struct device *dev, struct video_selection *sel);
 
+/**
+ * @brief Video driver API struct type.
+ */
 __subsystem struct video_driver_api {
-	/* mandatory callbacks */
+	/** Mandatory */
 	video_api_format_t set_format;
+	/** Mandatory */
 	video_api_format_t get_format;
+	/** Mandatory */
 	video_api_set_stream_t set_stream;
+	/** Mandatory */
 	video_api_get_caps_t get_caps;
-	/* optional callbacks */
+	/** Optional */
 	video_api_enqueue_t enqueue;
+	/** Optional */
 	video_api_dequeue_t dequeue;
+	/** Optional */
 	video_api_flush_t flush;
+	/** Optional */
 	video_api_ctrl_t set_ctrl;
+	/** Optional */
 	video_api_ctrl_t get_volatile_ctrl;
+	/** Optional */
 	video_api_set_signal_t set_signal;
+	/** Optional */
 	video_api_frmival_t set_frmival;
+	/** Optional */
 	video_api_frmival_t get_frmival;
+	/** Optional */
 	video_api_enum_frmival_t enum_frmival;
+	/** Optional */
 	video_api_selection_t set_selection;
+	/** Optional */
 	video_api_selection_t get_selection;
-	/** Transform capability from one end to the other end of an m2m device */
+	/** Optional */
 	video_api_transform_cap_t transform_cap;
 };
 
 /**
- * @brief Set video format.
+ * @brief Set video format of a driver.
  *
  * Configure video device with a specific format.
  *
  * @param dev Pointer to the device structure for the driver instance.
  * @param fmt Pointer to a video format struct.
  *
- * @retval 0 Is successful.
+ * @retval 0 If successful.
  * @retval -EINVAL If parameters are invalid.
  * @retval -ENOTSUP If format is not supported.
  * @retval -EIO General input / output error.
  */
-static inline int video_set_format(const struct device *dev, struct video_format *fmt)
+static inline int video_driver_set_format(const struct device *dev, struct video_format *fmt)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || fmt == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->set_format == NULL) {
 		return -ENOSYS;
 	}
@@ -448,7 +166,7 @@ static inline int video_set_format(const struct device *dev, struct video_format
 }
 
 /**
- * @brief Get video format.
+ * @brief Get video format of a driver.
  *
  * Get video device current video format.
  *
@@ -457,15 +175,10 @@ static inline int video_set_format(const struct device *dev, struct video_format
  *
  * @retval pointer to video format
  */
-static inline int video_get_format(const struct device *dev, struct video_format *fmt)
+static inline int video_driver_get_format(const struct device *dev, struct video_format *fmt)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || fmt == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->get_format == NULL) {
 		return -ENOSYS;
 	}
@@ -474,9 +187,7 @@ static inline int video_get_format(const struct device *dev, struct video_format
 }
 
 /**
- * @brief Set video frame interval.
- *
- * Configure video device with a specific frame interval.
+ * @brief Apply a video frame interval to a driver.
  *
  * Drivers must not return an error solely because the requested interval doesn’t match the device
  * capabilities. They must instead modify the interval to match what the hardware can provide.
@@ -489,19 +200,10 @@ static inline int video_get_format(const struct device *dev, struct video_format
  * @retval -EINVAL If parameters are invalid.
  * @retval -EIO General input / output error.
  */
-static inline int video_set_frmival(const struct device *dev, struct video_frmival *frmival)
+static inline int video_driver_set_frmival(const struct device *dev, struct video_frmival *frmival)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || frmival == NULL) {
-		return -EINVAL;
-	}
-
-	if (frmival->numerator == 0 || frmival->denominator == 0) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->set_frmival == NULL) {
 		return -ENOSYS;
 	}
@@ -510,7 +212,7 @@ static inline int video_set_frmival(const struct device *dev, struct video_frmiv
 }
 
 /**
- * @brief Get video frame interval.
+ * @brief Get video frame interval of a driver.
  *
  * Get current frame interval of the video device.
  *
@@ -522,15 +224,10 @@ static inline int video_set_frmival(const struct device *dev, struct video_frmiv
  * @retval -EINVAL If parameters are invalid.
  * @retval -EIO General input / output error.
  */
-static inline int video_get_frmival(const struct device *dev, struct video_frmival *frmival)
+static inline int video_driver_get_frmival(const struct device *dev, struct video_frmival *frmival)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || frmival == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->get_frmival == NULL) {
 		return -ENOSYS;
 	}
@@ -543,9 +240,9 @@ static inline int video_get_frmival(const struct device *dev, struct video_frmiv
  *
  * List all supported video frame intervals of a given format.
  *
- * Applications should fill the pixelformat, width and height fields of the
- * video_frmival_enum struct first to form a query. Then, the index field is
- * used to iterate through the supported frame intervals list.
+ * Drivers should fill the pixelformat, width and height fields of @ref video_frmival_enum given
+ * as query. The index field is used to iterate through the supported frame intervals for this
+ * format.
  *
  * @param dev Pointer to the device structure for the driver instance.
  * @param fie Pointer to a video frame interval enumeration struct.
@@ -555,15 +252,11 @@ static inline int video_get_frmival(const struct device *dev, struct video_frmiv
  * @retval -EINVAL If parameters are invalid.
  * @retval -EIO General input / output error.
  */
-static inline int video_enum_frmival(const struct device *dev, struct video_frmival_enum *fie)
+static inline int video_driver_enum_frmival(const struct device *dev,
+					    struct video_frmival_enum *fie)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || fie == NULL || fie->format == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->enum_frmival == NULL) {
 		return -ENOSYS;
 	}
@@ -572,20 +265,7 @@ static inline int video_enum_frmival(const struct device *dev, struct video_frmi
 }
 
 /**
- * @brief Import an external memory to the video buffer pool
- *
- * Import an externally allocated memory as a @ref video_buffer in the video buffer pool
- *
- * @param mem Pointer to the external memory
- * @param sz Size of the external memory
- * @param idx Returned index of the imported video buffer in the video buffer pool
- *
- * @retval 0 on success or a negative errno code on failure.
- */
-int video_import_buffer(uint8_t *mem, size_t sz, uint16_t *idx);
-
-/**
- * @brief Enqueue a video buffer.
+ * @brief Pass a video buffer to a driver.
  *
  * Enqueue an empty (capturing) or filled (output) video buffer in the driver’s
  * endpoint incoming queue.
@@ -593,14 +273,23 @@ int video_import_buffer(uint8_t *mem, size_t sz, uint16_t *idx);
  * @param dev Pointer to the device structure for the driver instance.
  * @param buf Pointer to the video buffer.
  *
- * @retval 0 Is successful.
+ * @retval 0 If successful.
  * @retval -EINVAL If parameters are invalid.
  * @retval -EIO General input / output error.
  */
-int video_enqueue(const struct device *dev, struct video_buffer *buf);
+static inline int video_driver_enqueue(const struct device *dev, struct video_buffer *vbuf)
+{
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
+
+	if (api->enqueue == NULL) {
+		return -ENOSYS;
+	}
+
+	return api->enqueue(dev, vbuf);
+}
 
 /**
- * @brief Dequeue a video buffer.
+ * @brief Dequeue a video buffer from a driver.
  *
  * Dequeue a filled (capturing) or displayed (output) buffer from the driver’s
  * endpoint outgoing queue.
@@ -609,29 +298,24 @@ int video_enqueue(const struct device *dev, struct video_buffer *buf);
  * @param buf Pointer a video buffer pointer.
  * @param timeout Timeout
  *
- * @retval 0 Is successful.
+ * @retval 0 If successful.
  * @retval -EINVAL If parameters are invalid.
  * @retval -EIO General input / output error.
  */
-static inline int video_dequeue(const struct device *dev, struct video_buffer **buf,
-				k_timeout_t timeout)
+static inline int video_driver_dequeue(const struct device *dev, struct video_buffer **vbuf,
+				       k_timeout_t timeout)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || buf == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->dequeue == NULL) {
 		return -ENOSYS;
 	}
 
-	return api->dequeue(dev, buf, timeout);
+	return api->dequeue(dev, vbuf, timeout);
 }
 
 /**
- * @brief Flush endpoint buffers.
+ * @brief Flush endpoint buffers from a driver.
  *
  * A call to flush finishes when all endpoint buffers have been moved from
  * incoming queue to outgoing queue. Either because canceled or fully processed
@@ -641,17 +325,12 @@ static inline int video_dequeue(const struct device *dev, struct video_buffer **
  * @param cancel If true, cancel buffer processing instead of waiting for
  *        completion.
  *
- * @retval 0 Is successful, -ERRNO code otherwise.
+ * @retval 0 If successful, -ERRNO code otherwise.
  */
-static inline int video_flush(const struct device *dev, bool cancel)
+static inline int video_driver_flush(const struct device *dev, bool cancel)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->flush == NULL) {
 		return -ENOSYS;
 	}
@@ -660,30 +339,22 @@ static inline int video_flush(const struct device *dev, bool cancel)
 }
 
 /**
- * @brief Start the video device function.
+ * @brief Start or stop the video driver function.
  *
- * video_stream_start is called to enter ‘streaming’ state (capture, output...).
- * The driver may receive buffers with video_enqueue() before video_stream_start
- * is called. If driver/device needs a minimum number of buffers before being
- * able to start streaming, then driver set the min_vbuf_count to the related
- * endpoint capabilities.
+ * Start (enable == true) or stop (enable == false) streaming on the video device.
  *
  * @param dev Pointer to the device structure.
- * @param type The type of the buffers stream to start.
+ * @param enable If true, start streaming, otherwise stop streaming.
+ * @param type The type of the buffers stream to start or stop.
  *
  * @retval 0 Successful.
- * @retval -EINVAL Parameters are invalid.
- * @retval -EIO General input / output error.
+ * @retval -ENOSYS API is not implemented.
  */
-static inline int video_stream_start(const struct device *dev, enum video_buf_type type)
+static inline int video_driver_set_stream(const struct device *dev, bool enable,
+					  enum video_buf_type type)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->set_stream == NULL) {
 		return -ENOSYS;
 	}
@@ -692,56 +363,64 @@ static inline int video_stream_start(const struct device *dev, enum video_buf_ty
 }
 
 /**
- * @brief Stop the video device function.
+ * @brief Get a volatile video control value of a driver.
  *
- * On video_stream_stop, driver must stop any transactions or wait until they
- * finish.
+ * Update the control value of @p cid by requesting it from the hardware.
+ * After this, the matching @c ctrl->val contains the updated value.
  *
  * @param dev Pointer to the device structure.
- * @param type The type of the buffers stream to stop.
+ * @param cid Id of the control to set/get its value.
  *
- * @retval 0 Successful.
- * @retval -EINVAL Parameters are invalid.
- * @retval -EIO General input / output error.
+ * @retval 0 If successful
+ * @retval -ENOSYS API is not implemented.
  */
-static inline int video_stream_stop(const struct device *dev, enum video_buf_type type)
+static inline int video_driver_get_volatile_ctrl(const struct device *dev, uint32_t cid)
 {
-	const struct video_driver_api *api;
-	int ret;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
-	if (api->set_stream == NULL) {
+	if (api->get_volatile_ctrl == NULL) {
 		return -ENOSYS;
 	}
 
-	ret = api->set_stream(dev, false, type);
-	video_flush(dev, true);
-
-	return ret;
+	return api->get_volatile_ctrl(dev, cid);
 }
 
 /**
- * @brief Get the capabilities of a video endpoint.
+ * @brief Set a video control value of a driver.
+ *
+ * Apply a video control value of @p cid to a device.
+ * The value is taken from the matching @c ctrl->val.
+ *
+ * @param dev Pointer to the device structure.
+ * @param cid Id of the control to set/get its value.
+ *
+ * @retval 0 If successful
+ * @retval -ENOSYS API is not implemented.
+ */
+static inline int video_driver_set_ctrl(const struct device *dev, uint32_t cid)
+{
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
+
+	if (api->set_ctrl == NULL) {
+		return -ENOSYS;
+	}
+
+	return api->set_ctrl(dev, cid);
+}
+
+/**
+ * @brief Get the capabilities of a video driver endpoint.
  *
  * @param dev Pointer to the device structure for the driver instance.
  * @param caps Pointer to the video_caps struct to fill.
  *
- * @retval 0 Is successful, -ERRNO code otherwise.
+ * @retval 0 If successful, -ERRNO code otherwise.
+ * @retval -ENOSYS API is not implemented.
  */
-static inline int video_get_caps(const struct device *dev, struct video_caps *caps)
+static inline int video_driver_get_caps(const struct device *dev, struct video_caps *caps)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || caps == NULL ||
-	    (caps->type != VIDEO_BUF_TYPE_INPUT && caps->type != VIDEO_BUF_TYPE_OUTPUT)) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->get_caps == NULL) {
 		return -ENOSYS;
 	}
@@ -752,24 +431,7 @@ static inline int video_get_caps(const struct device *dev, struct video_caps *ca
 /**
  * @brief Transform a video format capability from one end to the other end of a m2m video device.
  *
- * This function transforms a @ref video_format_cap from one end to the other end of an m2m video
- * device. It allows applications to iteratively get all the supported capabilities on one end of
- * the device given a single input capability on the other end.
- *
- * Applications pass the addresses of a single input cap and a single res_cap together with the
- * type of the res_cap and the index (started from 0) to iterate through to the function. The
- * driver fills the res_cap structure and return 0 each time. Index should be incremented to get
- * the next res_cap until the function returns an errno e.g., -EINVAL. For example:
- *
- * @code{.c}
- *  struct video_format_cap cap = {.pixelformat = VIDEO_PIX_FMT_RGB565};
- *  struct video_format_cap res_cap = {0};
- *  uint8_t ind = 0;
- *  while (video_transform_cap(dev, &cap, &res_cap, VIDEO_BUF_TYPE_OUTPUT, ind) == 0) {
- *		// Process output_cap here
- *		ind++;
- *  }
- * @endcode
+ * See @ref video_transform_cap for detailed description and usage.
  *
  * @param dev Pointer to the device structure.
  * @param cap Pointer to the source video format capability structure.
@@ -783,90 +445,19 @@ static inline int video_get_caps(const struct device *dev, struct video_caps *ca
  * @retval -ENOTSUP The transformation is not supported.
  * @retval -EIO General input / output error.
  */
-static inline int video_transform_cap(const struct device *const dev,
-				      const struct video_format_cap *const cap,
-				      struct video_format_cap *const res_cap,
-				      enum video_buf_type type, uint16_t ind)
+static inline int video_driver_transform_cap(const struct device *const dev,
+					     const struct video_format_cap *const cap,
+					     struct video_format_cap *const res_cap,
+					     enum video_buf_type type, uint16_t ind)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || cap == NULL || res_cap == NULL ||
-	    (type != VIDEO_BUF_TYPE_INPUT && type != VIDEO_BUF_TYPE_OUTPUT)) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->transform_cap == NULL) {
 		return -ENOSYS;
 	}
 
 	return api->transform_cap(dev, cap, res_cap, type, ind);
 }
-
-/**
- * @brief Set the value of a control.
- *
- * This set the value of a video control, value type depends on control ID, and
- * must be interpreted accordingly.
- *
- * @param dev Pointer to the device structure for the driver instance.
- * @param control Pointer to the video control struct.
- *
- * @retval 0 Is successful.
- * @retval -EINVAL If parameters are invalid.
- * @retval -ENOTSUP If format is not supported.
- * @retval -EIO General input / output error.
- */
-int video_set_ctrl(const struct device *dev, struct video_control *control);
-
-/**
- * @brief Get the current value of a control.
- *
- * This retrieve the value of a video control, value type depends on control ID,
- * and must be interpreted accordingly.
- *
- * @param dev Pointer to the device structure.
- * @param control Pointer to the video control struct.
- *
- * @retval 0 Is successful.
- * @retval -EINVAL If parameters are invalid.
- * @retval -ENOTSUP If format is not supported.
- * @retval -EIO General input / output error.
- */
-int video_get_ctrl(const struct device *dev, struct video_control *control);
-
-struct video_ctrl_query;
-
-/**
- * @brief Query information about a control.
- *
- * Applications set the id field of the query structure, the function fills the rest of this
- * structure. It is possible to enumerate base class controls (i.e., VIDEO_CID_BASE + x) by calling
- * this function with successive id values starting from VIDEO_CID_BASE up to and exclusive
- * VIDEO_CID_LASTP1. The function may return -ENOTSUP if a control in this range is not supported.
- * Applications can also enumerate private controls by starting at VIDEO_CID_PRIVATE_BASE and
- * incrementing the id until the driver returns -ENOTSUP. For other control classes, it's a bit more
- * difficult. Hence, the best way to enumerate all kinds of device's supported controls is to
- * iterate with VIDEO_CTRL_FLAG_NEXT_CTRL.
- *
- * @param cq Pointer to the control query struct.
- *
- * @retval 0 If successful.
- * @retval -EINVAL If the control id is invalid.
- * @retval -ENOTSUP If the control id is not supported.
- */
-int video_query_ctrl(struct video_ctrl_query *cq);
-
-/**
- * @brief Print all the information of a control.
- *
- * Print all the information of a control including its name, type, flag, range,
- * menu (if any) and current value, i.e. by invoking the video_get_ctrl(), in a
- * human readble format.
- *
- * @param cq Pointer to the control query struct.
- */
-void video_print_ctrl(const struct video_ctrl_query *const cq);
 
 /**
  * @brief Register/Unregister k_poll signal for a video endpoint.
@@ -878,17 +469,12 @@ void video_print_ctrl(const struct video_ctrl_query *const cq);
  * @param dev Pointer to the device structure for the driver instance.
  * @param sig Pointer to k_poll_signal
  *
- * @retval 0 Is successful, -ERRNO code otherwise.
+ * @retval 0 If successful, -ERRNO code otherwise.
  */
-static inline int video_set_signal(const struct device *dev, struct k_poll_signal *sig)
+static inline int video_driver_set_signal(const struct device *dev, struct k_poll_signal *sig)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || sig == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->set_signal == NULL) {
 		return -ENOSYS;
 	}
@@ -910,20 +496,15 @@ static inline int video_set_signal(const struct device *dev, struct k_poll_signa
  * @param dev Pointer to the device structure for the driver instance.
  * @param sel Pointer to a video selection structure
  *
- * @retval 0 Is successful.
+ * @retval 0 If successful.
  * @retval -EINVAL If parameters are invalid.
  * @retval -ENOTSUP If format is not supported.
  * @retval -EIO General input / output error.
  */
-static inline int video_set_selection(const struct device *dev, struct video_selection *sel)
+static inline int video_driver_set_selection(const struct device *dev, struct video_selection *sel)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || sel == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->set_selection == NULL) {
 		return -ENOSYS;
 	}
@@ -943,1510 +524,21 @@ static inline int video_set_selection(const struct device *dev, struct video_sel
  * @param dev Pointer to the device structure for the driver instance.
  * @param sel Pointer to a video selection structure, @c type and @c target set by the caller
  *
- * @retval 0 Is successful.
+ * @retval 0 If successful.
  * @retval -EINVAL If parameters are invalid.
  * @retval -ENOTSUP If format is not supported.
  * @retval -EIO General input / output error.
  */
-static inline int video_get_selection(const struct device *dev, struct video_selection *sel)
+static inline int video_driver_get_selection(const struct device *dev, struct video_selection *sel)
 {
-	const struct video_driver_api *api;
+	const struct video_driver_api *api = (const struct video_driver_api *)dev->api;
 
-	if (dev == NULL || sel == NULL) {
-		return -EINVAL;
-	}
-
-	api = (const struct video_driver_api *)dev->api;
 	if (api->get_selection == NULL) {
 		return -ENOSYS;
 	}
 
 	return api->get_selection(dev, sel);
 }
-
-/**
- * @brief Allocate aligned video buffer.
- *
- * @param size Size of the video buffer (in bytes).
- * @param align Alignment of the requested memory, must be a power of two.
- * @param timeout Timeout duration or K_NO_WAIT
- *
- * @return pointer to allocated video buffer
- */
-struct video_buffer *video_buffer_aligned_alloc(size_t size, size_t align, k_timeout_t timeout);
-
-/**
- * @brief Allocate video buffer.
- *
- * @param size Size of the video buffer (in bytes).
- * @param timeout Timeout duration or K_NO_WAIT
- *
- * @return pointer to allocated video buffer
- */
-struct video_buffer *video_buffer_alloc(size_t size, k_timeout_t timeout);
-
-/**
- * @brief Release a video buffer.
- *
- * @param buf Pointer to the video buffer to release.
- *
- * @retval 0 on success or a negative errno on failure
- */
-int video_buffer_release(struct video_buffer *buf);
-
-/**
- * @brief Search for a format that matches in a list of capabilities
- *
- * @param fmts The format capability list to search.
- * @param fmt The format to find in the list.
- * @param idx The pointer to a number of the first format that matches.
- *
- * @return 0 when a format is found.
- * @return -ENOENT when no matching format is found.
- */
-int video_format_caps_index(const struct video_format_cap *fmts, const struct video_format *fmt,
-			    size_t *idx);
-
-/**
- * @brief Compute the difference between two frame intervals
- *
- * @param frmival Frame interval to turn into microseconds.
- *
- * @return The frame interval value in microseconds.
- */
-static inline uint64_t video_frmival_nsec(const struct video_frmival *frmival)
-{
-	if (frmival == NULL || frmival->denominator == 0) {
-		return -EINVAL;
-	}
-
-	return (uint64_t)NSEC_PER_SEC * frmival->numerator / frmival->denominator;
-}
-
-/**
- * @brief Find the closest match to a frame interval value within a stepwise frame interval.
- *
- * @param stepwise The stepwise frame interval range to search
- * @param desired The frame interval for which find the closest match
- * @param match The resulting frame interval closest to @p desired
- */
-void video_closest_frmival_stepwise(const struct video_frmival_stepwise *stepwise,
-				    const struct video_frmival *desired,
-				    struct video_frmival *match);
-
-/**
- * @brief Find the closest match to a frame interval value within a video device.
- *
- * To compute the closest match, fill @p match with the following fields:
- *
- * - @c match->format to the @ref video_format of interest.
- * - @c match->type to @ref VIDEO_FRMIVAL_TYPE_DISCRETE.
- * - @c match->discrete to the desired frame interval.
- *
- * The result will be loaded into @p match, with the following fields set:
- *
- * - @c match->discrete to the value of the closest frame interval.
- * - @c match->index to the index of the closest frame interval.
- *
- * @param dev Video device to query.
- * @param match Frame interval enumerator with the query, and loaded with the result.
- */
-void video_closest_frmival(const struct device *dev, struct video_frmival_enum *match);
-
-/**
- * @brief Return the link-frequency advertised by a device
- *
- * Device exposing a CSI link should advertise at least one of the following two controls:
- *   - @ref VIDEO_CID_LINK_FREQ
- *   - @ref VIDEO_CID_PIXEL_RATE
- *
- * At first the helper will try read the @ref VIDEO_CID_LINK_FREQ and if not available will
- * approximate the link-frequency from the @ref VIDEO_CID_PIXEL_RATE value, taking into
- * consideration the bits per pixel of the format and the number of lanes.
- *
- * @param dev Video device to query.
- * @param bpp Amount of bits per pixel of the pixel format produced by the device
- * @param lane_nb Number of CSI-2 lanes used
- */
-int64_t video_get_csi_link_freq(const struct device *dev, uint8_t bpp, uint8_t lane_nb);
-
-/**
- * @brief Estimate the size and pitch in bytes of a @ref video_format
- *
- * This helper should only be used by drivers that support the whole image frame.
- *
- * For uncompressed formats, it gives the actual size and pitch of the
- * whole raw image without any padding.
- *
- * For compressed formats, it gives a rough estimate size of a complete
- * compressed frame.
- *
- * @param fmt Pointer to the video format structure
- * @return 0 on success, otherwise a negative errno code
- */
-int video_estimate_fmt_size(struct video_format *fmt);
-
-/**
- * @brief Set compose rectangle (if applicable) prior to setting format
- *
- * Some devices expose compose capabilities, allowing them to apply a transformation
- * (downscale / upscale) to the frame. For those devices, it is necessary to set the
- * compose rectangle before being able to apply the frame format (which must have the
- * same width / height as the compose rectangle width / height).
- * In order to allow non-compose aware application to be able to control such devices,
- * introduce a helper which, if available, will apply the compose rectangle prior to
- * setting the format.
- *
- * @param dev Pointer to the video device struct to set format
- * @param fmt Pointer to a video format struct.
- *
- * @retval 0 Is successful.
- * @retval -EINVAL If parameters are invalid.
- * @retval -ENOTSUP If format is not supported.
- * @retval -EIO General input / output error.
- */
-int video_set_compose_format(const struct device *dev, struct video_format *fmt);
-
-/**
- * @brief Transfer a buffer between 2 video device
- *
- * Helper function which dequeues a buffer from a source device and enqueues it into a
- * sink device, changing its buffer type between the two.
- *
- * @param src		Video device from where buffer is dequeued (source)
- * @param sink		Video device into which the buffer is queued (sink)
- * @param src_type	Video buffer type on the source device
- * @param sink_type	Video buffer type on the sink device
- * @param timeout	Timeout to be applied on dequeue
- *
- * @return 0 on success, otherwise a negative errno code
- */
-int video_transfer_buffer(const struct device *src, const struct device *sink,
-			  enum video_buf_type src_type, enum video_buf_type sink_type,
-			  k_timeout_t timeout);
-
-/**
- * @defgroup video_pixel_formats Video pixel formats
- * The '|' characters separate the pixels or logical blocks, and spaces separate the bytes.
- * The uppercase letter represents the most significant bit.
- * The lowercase letters represent the rest of the bits.
- * @{
- */
-
-/**
- * @brief Four-character-code uniquely identifying the pixel format
- */
-#define VIDEO_FOURCC(a, b, c, d)                                                                   \
-	((uint32_t)(a) | ((uint32_t)(b) << 8) | ((uint32_t)(c) << 16) | ((uint32_t)(d) << 24))
-
-/**
- * @brief Convert a four-character-string to a four-character-code
- *
- * Convert a string literal or variable into a four-character-code
- * as defined by @ref VIDEO_FOURCC.
- *
- * @param str String to be converted
- * @return Four-character-code.
- */
-#define VIDEO_FOURCC_FROM_STR(str) VIDEO_FOURCC((str)[0], (str)[1], (str)[2], (str)[3])
-
-/**
- * @brief Convert a four-character-code to a four-character-string
- *
- * Convert a four-character code as defined by @ref VIDEO_FOURCC into a string that can be used
- * anywhere, such as in debug logs with the %s print formatter.
- *
- * @param fourcc The 32-bit four-character-code integer to be converted, in CPU-native endinaness.
- * @return Four-character-string built out of it.
- */
-#define VIDEO_FOURCC_TO_STR(fourcc)                                                                \
-	((char[]){                                                                                 \
-		(char)((fourcc) & 0xFF),                                                           \
-		(char)(((fourcc) >> 8) & 0xFF),                                                    \
-		(char)(((fourcc) >> 16) & 0xFF),                                                   \
-		(char)(((fourcc) >> 24) & 0xFF),                                                   \
-		'\0'                                                                               \
-	})
-
-/**
- * @name Bayer formats (R, G, B channels).
- *
- * The full color information is spread over multiple pixels.
- *
- * When the format includes more than 8-bit per pixel, a strategy becomes needed to pack
- * the bits over multiple bytes, as illustrated for each format.
- *
- * The number above the 'R', 'r', 'G', 'g', 'B', 'b' are hints about which pixel number the
- * following bits belong to.
- *
- * @{
- */
-
-/**
- * @brief Repeat a macro for every Bayer format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_BAYER(X, ...)						\
-	VIDEO_FOREACH_BAYER_PADDED(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_BAYER_MIPI_PACKED(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_BAYER_NON_PACKED(X, __VA_ARGS__)
-
-/**
- * @brief Repeat a macro for every Bayer padded format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_BAYER_PADDED(X, ...)					\
-	X(VIDEO_PIX_FMT_SBGGR8P16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG8P16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG8P16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB8P16, __VA_ARGS__)
-
-/**
- * 8-bit bayer format, split in two 4-bit blocks each zero-padded, little endian.
- *
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | 0000bbbb 0000Bbbb | 0000gggg 0000Gggg | 0000gggg 0000Gggg | 0000rrrr 0000Rrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR8P16 VIDEO_FOURCC('p', 'B', '8', '2')
-
-/**
- * 8-bit bayer format, split in two 4-bit blocks each zero-padded, little endian.
- *
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | 0000gggg 0000Gggg | 0000bbbb 0000Bbbb | 0000rrrr 0000Rrrr | 0000gggg 0000Gggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG8P16 VIDEO_FOURCC('p', 'G', '8', '2')
-
-/**
- * 8-bit bayer format, split in two 4-bit blocks each zero-padded, little endian.
- *
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | 0000gggg 0000Gggg | 0000rrrr 0000Rrrr | 0000bbbb 0000Bbbb | 0000gggg 0000Gggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG8P16 VIDEO_FOURCC('p', 'g', '8', '2')
-
-/**
- * 8-bit bayer format, split in two 4-bit blocks each zero-padded, little endian.
- *
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | 0000rrrr 0000Rrrr | 0000gggg 0000Gggg | 0000gggg 0000Gggg | 0000bbbb 0000Bbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB8P16 VIDEO_FOURCC('p', 'R', '8', '2')
-
-/**
- * @brief Repeat a macro for every Bayer MIPI-packed format, passed as first parameter
- *
- * A red, green, blue channel for every pixel, other than 8-bit per pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_BAYER_MIPI_PACKED(X, ...)					\
-	X(VIDEO_PIX_FMT_SBGGR10P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG10P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG10P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB10P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SBGGR12P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG12P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG12P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB12P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SBGGR14P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG14P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG14P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB14P, __VA_ARGS__)
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          3 2 1 0
- * | Bbbbbbbb | Gggggggg | Bbbbbbbb | Gggggggg | ggbbggbb | ...
- * | Gggggggg | Rrrrrrrr | Gggggggg | Rrrrrrrr | rrggrrgg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR10P VIDEO_FOURCC('p', 'B', 'A', 'A')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          3 2 1 0
- * | Gggggggg | Bbbbbbbb | Gggggggg | Bbbbbbbb | bbggbbgg | ...
- * | Rrrrrrrr | Gggggggg | Rrrrrrrr | Gggggggg | ggrrggrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG10P VIDEO_FOURCC('p', 'G', 'A', 'A')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          3 2 1 0
- * | Gggggggg | Rrrrrrrr | Gggggggg | Rrrrrrrr | rrggrrgg | ...
- * | Bbbbbbbb | Gggggggg | Bbbbbbbb | Gggggggg | ggbbggbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG10P VIDEO_FOURCC('p', 'g', 'A', 'A')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          3 2 1 0
- * | Rrrrrrrr | Gggggggg | Rrrrrrrr | Gggggggg | ggrrggrr | ...
- * | Gggggggg | Bbbbbbbb | Gggggggg | Bbbbbbbb | bbggbbgg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB10P VIDEO_FOURCC('p', 'R', 'A', 'A')
-
-/**
- * @code{.unparsed}
- *   0          1          1   0      2          3          3   2
- * | Bbbbbbbb | Gggggggg | ggggbbbb | Bbbbbbbb | Gggggggg | ggggbbbb | ...
- * | Gggggggg | Rrrrrrrr | rrrrgggg | Gggggggg | Rrrrrrrr | rrrrgggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR12P VIDEO_FOURCC('p', 'B', 'C', 'C')
-
-/**
- * @code{.unparsed}
- *   0          1          1   0      2          3          3   2
- * | Gggggggg | Bbbbbbbb | bbbbgggg | Gggggggg | Bbbbbbbb | bbbbgggg | ...
- * | Rrrrrrrr | Gggggggg | ggggrrrr | Rrrrrrrr | Gggggggg | ggggrrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG12P VIDEO_FOURCC('p', 'G', 'C', 'C')
-
-/**
- * @code{.unparsed}
- *   0          1          1   0      2          3          3   2
- * | Gggggggg | Rrrrrrrr | rrrrgggg | Gggggggg | Rrrrrrrr | rrrrgggg | ...
- * | Bbbbbbbb | Gggggggg | ggggbbbb | Bbbbbbbb | Gggggggg | ggggbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG12P VIDEO_FOURCC('p', 'g', 'C', 'C')
-
-/**
- * @code{.unparsed}
- *   0          1          1   0      2          3          3   2
- * | Rrrrrrrr | Gggggggg | ggggrrrr | Rrrrrrrr | Gggggggg | ggggrrrr | ...
- * | Gggggggg | Bbbbbbbb | bbbbgggg | Gggggggg | Bbbbbbbb | bbbbgggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB12P VIDEO_FOURCC('p', 'R', 'C', 'C')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          1 0      2   1    3     2
- * | Bbbbbbbb | Gggggggg | Bbbbbbbb | Gggggggg | ggbbbbbb bbbbgggg ggggggbb | ...
- * | Gggggggg | Rrrrrrrr | Gggggggg | Rrrrrrrr | rrgggggg ggggrrrr rrrrrrgg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR14P VIDEO_FOURCC('p', 'B', 'E', 'E')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          1 0      2   1    3     2
- * | Gggggggg | Bbbbbbbb | Gggggggg | Bbbbbbbb | bbgggggg ggggbbbb bbbbbbgg | ...
- * | Rrrrrrrr | Gggggggg | Rrrrrrrr | Gggggggg | ggrrrrrr rrrrgggg ggggggrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG14P VIDEO_FOURCC('p', 'G', 'E', 'E')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          1 0      2   1    3     2
- * | Gggggggg | Rrrrrrrr | Gggggggg | Rrrrrrrr | rrgggggg ggggrrrr rrrrrrgg | ...
- * | Bbbbbbbb | Gggggggg | Bbbbbbbb | Gggggggg | ggbbbbbb bbbbgggg ggggggbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG14P VIDEO_FOURCC('p', 'g', 'E', 'E')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          1 0      2   1    3     2
- * | Rrrrrrrr | Gggggggg | Rrrrrrrr | Gggggggg | ggrrrrrr rrrrgggg ggggggrr | ...
- * | Gggggggg | Bbbbbbbb | Gggggggg | Bbbbbbbb | bbgggggg ggggbbbb bbbbbbgg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB14P VIDEO_FOURCC('p', 'R', 'E', 'E')
-
-/**
- * @brief Repeat a macro for every Bayer non-packed format, passed as first parameter
- *
- * A red, green, blue channel for every pixel, other than 8-bit per pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_BAYER_NON_PACKED(X, ...)					\
-	X(VIDEO_PIX_FMT_SBGGR8, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG8, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG8, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB8, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG10, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG10, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB10, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SBGGR12, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG12, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG12, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB12, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SBGGR14, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG14, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG14, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB14, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SBGGR16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGBRG16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SGRBG16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_SRGGB16, __VA_ARGS__)
-
-/**
- * @code{.unparsed}
- *   0          1          2          3
- * | Bbbbbbbb | Gggggggg | Bbbbbbbb | Gggggggg | ...
- * | Gggggggg | Rrrrrrrr | Gggggggg | Rrrrrrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR8 VIDEO_FOURCC('B', 'A', '8', '1')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3
- * | Gggggggg | Bbbbbbbb | Gggggggg | Bbbbbbbb | ...
- * | Rrrrrrrr | Gggggggg | Rrrrrrrr | Gggggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG8 VIDEO_FOURCC('G', 'B', 'R', 'G')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3
- * | Gggggggg | Rrrrrrrr | Gggggggg | Rrrrrrrr | ...
- * | Bbbbbbbb | Gggggggg | Bbbbbbbb | Gggggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG8 VIDEO_FOURCC('G', 'R', 'B', 'G')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3
- * | Rrrrrrrr | Gggggggg | Rrrrrrrr | Gggggggg | ...
- * | Gggggggg | Bbbbbbbb | Gggggggg | Bbbbbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB8 VIDEO_FOURCC('R', 'G', 'G', 'B')
-
-/**
- * @code{.unparsed}
- * | bbbbbbbb 000000Bb | gggggggg 000000Gg | bbbbbbbb 000000Bb | gggggggg 000000Gg | ...
- * | gggggggg 000000Gg | rrrrrrrr 000000Rr | gggggggg 000000Gg | rrrrrrrr 000000Rr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR10 VIDEO_FOURCC('B', 'G', '1', '0')
-
-/**
- * @code{.unparsed}
- * | gggggggg 000000Gg | bbbbbbbb 000000Bb | gggggggg 000000Gg | bbbbbbbb 000000Bb | ...
- * | rrrrrrrr 000000Rr | gggggggg 000000Gg | rrrrrrrr 000000Rr | gggggggg 000000Gg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG10 VIDEO_FOURCC('G', 'B', '1', '0')
-
-/**
- * @code{.unparsed}
- * | gggggggg 000000Gg | rrrrrrrr 000000Rr | gggggggg 000000Gg | rrrrrrrr 000000Rr | ...
- * | bbbbbbbb 000000Bb | gggggggg 000000Gg | bbbbbbbb 000000Bb | gggggggg 000000Gg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG10 VIDEO_FOURCC('B', 'A', '1', '0')
-
-/**
- * @code{.unparsed}
- * | rrrrrrrr 000000Rr | gggggggg 000000Gg | rrrrrrrr 000000Rr | gggggggg 000000Gg | ...
- * | gggggggg 000000Gg | bbbbbbbb 000000Bb | gggggggg 000000Gg | bbbbbbbb 000000Bb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB10 VIDEO_FOURCC('R', 'G', '1', '0')
-
-/**
- * @code{.unparsed}
- * | bbbbbbbb 0000Bbbb | gggggggg 0000Gggg | bbbbbbbb 0000Bbbb | gggggggg 0000Gggg | ...
- * | gggggggg 0000Gggg | rrrrrrrr 0000Rrrr | gggggggg 0000Gggg | rrrrrrrr 0000Rrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR12 VIDEO_FOURCC('B', 'G', '1', '2')
-
-/**
- * @code{.unparsed}
- * | gggggggg 0000Gggg | bbbbbbbb 0000Bbbb | gggggggg 0000Gggg | bbbbbbbb 0000Bbbb | ...
- * | rrrrrrrr 0000Rrrr | gggggggg 0000Gggg | rrrrrrrr 0000Rrrr | gggggggg 0000Gggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG12 VIDEO_FOURCC('G', 'B', '1', '2')
-
-/**
- * @code{.unparsed}
- * | gggggggg 0000Gggg | rrrrrrrr 0000Rrrr | gggggggg 0000Gggg | rrrrrrrr 0000Rrrr | ...
- * | bbbbbbbb 0000Bbbb | gggggggg 0000Gggg | bbbbbbbb 0000Bbbb | gggggggg 0000Gggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG12 VIDEO_FOURCC('B', 'A', '1', '2')
-
-/**
- * @code{.unparsed}
- * | rrrrrrrr 0000Rrrr | gggggggg 0000Gggg | rrrrrrrr 0000Rrrr | gggggggg 0000Gggg | ...
- * | gggggggg 0000Gggg | bbbbbbbb 0000Bbbb | gggggggg 0000Gggg | bbbbbbbb 0000Bbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB12 VIDEO_FOURCC('R', 'G', '1', '2')
-
-/**
- * @code{.unparsed}
- * | bbbbbbbb 00Bbbbbb | gggggggg 00Gggggg | bbbbbbbb 00Bbbbbb | gggggggg 00Gggggg | ...
- * | gggggggg 00Gggggg | rrrrrrrr 00Rrrrrr | gggggggg 00Gggggg | rrrrrrrr 00Rrrrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR14 VIDEO_FOURCC('B', 'G', '1', '4')
-
-/**
- * @code{.unparsed}
- * | gggggggg 00Gggggg | bbbbbbbb 00Bbbbbb | gggggggg 00Gggggg | bbbbbbbb 00Bbbbbb | ...
- * | rrrrrrrr 00Rrrrrr | gggggggg 00Gggggg | rrrrrrrr 00Rrrrrr | gggggggg 00Gggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG14 VIDEO_FOURCC('G', 'B', '1', '4')
-
-/**
- * @code{.unparsed}
- * | gggggggg 00Gggggg | rrrrrrrr 00Rrrrrr | gggggggg 00Gggggg | rrrrrrrr 00Rrrrrr | ...
- * | bbbbbbbb 00Bbbbbb | gggggggg 00Gggggg | bbbbbbbb 00Bbbbbb | gggggggg 00Gggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG14 VIDEO_FOURCC('G', 'R', '1', '4')
-
-/**
- * @code{.unparsed}
- * | rrrrrrrr 00Rrrrrr | gggggggg 00Gggggg | rrrrrrrr 00Rrrrrr | gggggggg 00Gggggg | ...
- * | gggggggg 00Gggggg | bbbbbbbb 00Bbbbbb | gggggggg 00Gggggg | bbbbbbbb 00Bbbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB14 VIDEO_FOURCC('R', 'G', '1', '4')
-
-/**
- * @code{.unparsed}
- * | bbbbbbbb Bbbbbbbb | gggggggg Gggggggg | bbbbbbbb Bbbbbbbb | gggggggg Gggggggg | ...
- * | gggggggg Gggggggg | rrrrrrrr Rrrrrrrr | gggggggg Gggggggg | rrrrrrrr Rrrrrrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SBGGR16 VIDEO_FOURCC('B', 'Y', 'R', '2')
-
-/**
- * @code{.unparsed}
- * | gggggggg Gggggggg | bbbbbbbb Bbbbbbbb | gggggggg Gggggggg | bbbbbbbb Bbbbbbbb | ...
- * | rrrrrrrr Rrrrrrrr | gggggggg Gggggggg | rrrrrrrr Rrrrrrrr | gggggggg Gggggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGBRG16 VIDEO_FOURCC('G', 'B', '1', '6')
-
-/**
- * @code{.unparsed}
- * | gggggggg Gggggggg | rrrrrrrr Rrrrrrrr | gggggggg Gggggggg | rrrrrrrr Rrrrrrrr | ...
- * | bbbbbbbb Bbbbbbbb | gggggggg Gggggggg | bbbbbbbb Bbbbbbbb | gggggggg Gggggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SGRBG16 VIDEO_FOURCC('G', 'R', '1', '6')
-
-/**
- * @code{.unparsed}
- * | rrrrrrrr Rrrrrrrr | gggggggg Gggggggg | rrrrrrrr Rrrrrrrr | gggggggg Gggggggg | ...
- * | gggggggg Gggggggg | bbbbbbbb Bbbbbbbb | gggggggg Gggggggg | bbbbbbbb Bbbbbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_SRGGB16 VIDEO_FOURCC('R', 'G', '1', '6')
-
-/**
- * @}
- */
-
-/**
- * @name Grayscale formats
- * Luminance (Y) channel only, in various bit depth and packing.
- *
- * When the format includes more than 8-bit per pixel, a strategy becomes needed to pack
- * the bits over multiple bytes, as illustrated for each format.
- *
- * The number above the 'Y', 'y' are hints about which pixel number the following bits belong to.
- *
- * @{
- */
-
-/**
- * @brief Repeat a macro for every grayscale format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_GRAYSCALE(X, ...)						\
-	VIDEO_FOREACH_GRAYSCALE_NON_PACKED(X, __VA_ARGS__)			\
-	VIDEO_FOREACH_GRAYSCALE_PADDED(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_GRAYSCALE_MIPI_PACKED(X, __VA_ARGS__)
-
-/**
- * @brief Repeat a macro for every grayscale non-packed format, passed as first parameter
- *
- * A red, green, blue channel for every pixel, other than 8-bit per pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_GRAYSCALE_NON_PACKED(X, ...)				\
-	X(VIDEO_PIX_FMT_GREY, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_Y10, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_Y12, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_Y14, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_Y16, __VA_ARGS__)
-
-/**
- * Same as Y8 (8-bit luma-only) following the standard FOURCC naming,
- * or L8 in some graphics libraries.
- *
- * @code{.unparsed}
- *   0          1          2          3
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_GREY VIDEO_FOURCC('G', 'R', 'E', 'Y')
-
-/**
- * Little endian, with the 6 most significant bits set to Zero.
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | yyyyyyyy 000000Yy | yyyyyyyy 000000Yy | yyyyyyyy 000000Yy | yyyyyyyy 000000Yy | ...
- * | yyyyyyyy 000000Yy | yyyyyyyy 000000Yy | yyyyyyyy 000000Yy | yyyyyyyy 000000Yy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y10 VIDEO_FOURCC('Y', '1', '0', ' ')
-
-/**
- * Little endian, with the 4 most significant bits set to Zero.
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | yyyyyyyy 0000Yyyy | yyyyyyyy 0000Yyyy | yyyyyyyy 0000Yyyy | yyyyyyyy 0000Yyyy | ...
- * | yyyyyyyy 0000Yyyy | yyyyyyyy 0000Yyyy | yyyyyyyy 0000Yyyy | yyyyyyyy 0000Yyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y12 VIDEO_FOURCC('Y', '1', '2', ' ')
-
-/**
- * Little endian, with the 2 most significant bits set to Zero.
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | yyyyyyyy 00Yyyyyy | yyyyyyyy 00Yyyyyy | yyyyyyyy 00Yyyyyy | yyyyyyyy 00Yyyyyy | ...
- * | yyyyyyyy 00Yyyyyy | yyyyyyyy 00Yyyyyy | yyyyyyyy 00Yyyyyy | yyyyyyyy 00Yyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y14 VIDEO_FOURCC('Y', '1', '4', ' ')
-
-/**
- * Little endian.
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | yyyyyyyy Yyyyyyyy | yyyyyyyy Yyyyyyyy | yyyyyyyy Yyyyyyyy | yyyyyyyy Yyyyyyyy | ...
- * | yyyyyyyy Yyyyyyyy | yyyyyyyy Yyyyyyyy | yyyyyyyy Yyyyyyyy | yyyyyyyy Yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y16 VIDEO_FOURCC('Y', '1', '6', ' ')
-
-/**
- * @brief Repeat a macro for every grayscale padded format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_GRAYSCALE_PADDED(X, ...)					\
-	X(VIDEO_PIX_FMT_Y8P16, __VA_ARGS__)
-
-/**
- * 8-bit luma-only split in two 4-bit blocks each zero-padded, little endian.
- *
- * @code{.unparsed}
- *   0                   1                   2                   3
- * | 0000yyyy 0000Yyyy | 0000yyyy 0000Yyyy | 0000yyyy 0000Yyyy | 0000yyyy 0000Yyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y8P16 VIDEO_FOURCC('Y', '8', 'P', '2')
-
-/**
- * @brief Repeat a macro for every grayscale MIPI-packed format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_GRAYSCALE_MIPI_PACKED(X, ...)				\
-	X(VIDEO_PIX_FMT_Y10P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_Y12P, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_Y14P, __VA_ARGS__)
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          3 2 1 0
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y10P VIDEO_FOURCC('Y', '1', '0', 'P')
-
-/**
- * @code{.unparsed}
- *   0          1          1   0      2          3          3   2
- * | Yyyyyyyy | Yyyyyyyy | yyyyyyyy | Yyyyyyyy | Yyyyyyyy | yyyyyyyy | ...
- * | Yyyyyyyy | Yyyyyyyy | yyyyyyyy | Yyyyyyyy | Yyyyyyyy | yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y12P VIDEO_FOURCC('Y', '1', '2', 'P')
-
-/**
- * @code{.unparsed}
- *   0          1          2          3          1 0      2   1    3     2
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | yyyyyyyy yyyyyyyy yyyyyyyy | ...
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | yyyyyyyy yyyyyyyy yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_Y14P VIDEO_FOURCC('Y', '1', '4', 'P')
-
-/**
- * @}
- */
-
-/**
- * @name RGB formats
- * Per-color (R, G, B) channels.
- * @{
- */
-
-/**
- * @brief Repeat a macro for every RGB format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_RGB(X, ...)						\
-	VIDEO_FOREACH_RGB_PACKED(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_RGB_NON_PACKED(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_RGB_ALPHA(X, __VA_ARGS__)					\
-	VIDEO_FOREACH_RGB_PADDED(X, __VA_ARGS__)
-
-/**
- * @brief Repeat a macro for every RGB packed format, passed as first parameter
- *
- * A red, green, blue channel for every pixel, other than 8-bit per pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_RGB_PACKED(X, ...)					\
-	X(VIDEO_PIX_FMT_RGB565X, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_RGB565, __VA_ARGS__)
-
-/**
- * 5 red bits [15:11], 6 green bits [10:5], 5 blue bits [4:0].
- * This 16-bit integer is then packed in big endian format over two bytes:
- *
- * @code{.unparsed}
- *   15.....8 7......0
- * | RrrrrGgg gggBbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_RGB565X VIDEO_FOURCC('R', 'G', 'B', 'R')
-
-/**
- * 5 red bits [15:11], 6 green bits [10:5], 5 blue bits [4:0].
- * This 16-bit integer is then packed in little endian format over two bytes:
- *
- * @code{.unparsed}
- *   7......0 15.....8
- * | gggBbbbb RrrrrGgg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_RGB565 VIDEO_FOURCC('R', 'G', 'B', 'P')
-
-/**
- * @brief Repeat a macro for every RGB non-packed format, passed as first parameter
- *
- * A red, green, blue channel for every pixel, 8-bit per pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_RGB_NON_PACKED(X, ...)					\
-	X(VIDEO_PIX_FMT_BGR24, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_RGB24, __VA_ARGS__)
-
-/**
- * 24 bit RGB format with 8 bit per component
- *
- * @code{.unparsed}
- * | Bbbbbbbb Gggggggg Rggggggg | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_BGR24 VIDEO_FOURCC('B', 'G', 'R', '3')
-
-/**
- * 24 bit RGB format with 8 bit per component
- *
- * @code{.unparsed}
- * | Rggggggg Gggggggg Bbbbbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_RGB24 VIDEO_FOURCC('R', 'G', 'B', '3')
-
-/**
- * @brief Repeat a macro for every RGB alpha format, passed as first parameter
- *
- * A red, green, blue, alpha channel for every pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_RGB_ALPHA(X, ...)						\
-	X(VIDEO_PIX_FMT_ARGB32, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_ABGR32, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_RGBA32, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_BGRA32, __VA_ARGS__)
-
-/**
- * The first byte is alpha (A) for each pixel.
- *
- * @code{.unparsed}
- * | Aaaaaaaa Rrrrrrrr Gggggggg Bbbbbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_ARGB32 VIDEO_FOURCC('B', 'A', '2', '4')
-
-/**
- * The last byte is alpha (A) for each pixel.
- * *
- * @warning Linux calls this format ABGR32 due to a historical typo
- *
- * @code{.unparsed}
- * | Bbbbbbbb Gggggggg Rrrrrrrr Aaaaaaaa | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_BGRA32 VIDEO_FOURCC('A', 'R', '2', '4')
-
-/**
- * The last byte is alpha (A) for each pixel.
- *
- * @code{.unparsed}
- * | Rrrrrrrr Gggggggg Bbbbbbbb Aaaaaaaa | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_RGBA32 VIDEO_FOURCC('A', 'B', '2', '4')
-
-/**
- * The first byte is alpha (A) for each pixel.
- *
- * @warning Linux calls this format BGRA32 due to a historical typo
- *
- * @code{.unparsed}
- * | Aaaaaaaa Bbbbbbbb Gggggggg Rrrrrrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_ABGR32 VIDEO_FOURCC('R', 'A', '2', '4')
-
-/**
- * @brief Repeat a macro for every RGB padded format, passed as first parameter
- *
- * A red, green, blue channel for every pixel.
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_RGB_PADDED(X, ...)					\
-	X(VIDEO_PIX_FMT_XRGB32, __VA_ARGS__)
-
-/**
- * The first byte is empty (X) for each pixel.
- *
- * @code{.unparsed}
- * | Xxxxxxxx Rrrrrrrr Gggggggg Bbbbbbbb | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_XRGB32 VIDEO_FOURCC('B', 'X', '2', '4')
-
-/**
- * The first byte is empty (X) for each pixel.
- *
- * @warning Linux calls this format BGRX32 due to a historical typo
- *
- * @code{.unparsed}
- * | Xxxxxxxx Bbbbbbbb Gggggggg Rrrrrrrr | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_XBGR32 VIDEO_FOURCC('R', 'X', '2', '4')
-
-/**
- * The last byte is empty (X) for each pixel.
- *
- * @warning Linux calls this format XBGR32 due to a historical typo
- *
- * @code{.unparsed}
- * | Bbbbbbbb Gggggggg Rrrrrrrr Xxxxxxxx | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_BGRX32 VIDEO_FOURCC('X', 'R', '2', '4')
-
-/**
- * The last byte is empty (X) for each pixel.
- *
- * @code{.unparsed}
- * | Rrrrrrrr Gggggggg Bbbbbbbb Xxxxxxxx | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_RGBX32 VIDEO_FOURCC('X', 'B', '2', '4')
-
-/**
- * @}
- */
-
-/**
- * @name YUV formats
- * Luminance (Y) and chrominance (U, V) channels.
- * @{
- */
-
-/**
- * @brief Repeat a macro for every YUV format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_YUV(X, ...)						\
-	VIDEO_FOREACH_YUV_FULL_PLANAR(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_YUV_NON_PLANAR(X, __VA_ARGS__)				\
-	VIDEO_FOREACH_YUV_SEMI_PLANAR(X, __VA_ARGS__)
-
-/**
- * @brief Repeat a macro for every YUV Non-Planar format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_YUV_NON_PLANAR(X, ...)					\
-	X(VIDEO_PIX_FMT_YUYV, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_YVYU, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_VYUY, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_UYVY, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_YUV24, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_XYUV32, __VA_ARGS__)
-
-/**
- * There is either a missing channel per pixel, U or V.
- * The value is to be averaged over 2 pixels to get the value of individual pixel.
- *
- * @code{.unparsed}
- * | Yyyyyyyy Uuuuuuuu | Yyyyyyyy Vvvvvvvv | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_YUYV VIDEO_FOURCC('Y', 'U', 'Y', 'V')
-
-/**
- * @code{.unparsed}
- * | Yyyyyyyy Vvvvvvvv | Yyyyyyyy Uuuuuuuu | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_YVYU VIDEO_FOURCC('Y', 'V', 'Y', 'U')
-
-/**
- * @code{.unparsed}
- * | Vvvvvvvv Yyyyyyyy | Uuuuuuuu Yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_VYUY VIDEO_FOURCC('V', 'Y', 'U', 'Y')
-
-/**
- * @code{.unparsed}
- * | Uuuuuuuu Yyyyyyyy | Vvvvvvvv Yyyyyyyy | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_UYVY VIDEO_FOURCC('U', 'Y', 'V', 'Y')
-
-/**
- * The first byte is empty (X) for each pixel.
- *
- * @code{.unparsed}
- * | Xxxxxxxx Yyyyyyyy Uuuuuuuu Vvvvvvvv | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_XYUV32 VIDEO_FOURCC('X', 'Y', 'U', 'V')
-
-/**
- * 24 bit YUV format with 8 bit per component
- *
- * @code{.unparsed}
- * | Yyyyyyyy Uuuuuuuu Vvvvvvvv | ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_YUV24 VIDEO_FOURCC('Y', 'U', 'V', '3')
-
-/**
- * @brief Repeat a macro for every YUV Semi-Planar format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_YUV_SEMI_PLANAR(X, ...)					\
-	X(VIDEO_PIX_FMT_NV12, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_NV21, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_NV16, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_NV61, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_NV24, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_NV42, __VA_ARGS__)
-
-/**
- * Chroma (U/V) are subsampled horizontaly and vertically
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | ... |
- * | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | ...
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  U0/1/6/7  V0/1/6/7  U2/3/8/9  V2/3/8/9  ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_NV12 VIDEO_FOURCC('N', 'V', '1', '2')
-
-/**
- * Chroma (U/V) are subsampled horizontaly and vertically
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | ... |
- * | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | ...
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  V0/1/6/7  U0/1/6/7  V2/3/8/9  U2/3/8/9  ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_NV21 VIDEO_FOURCC('N', 'V', '2', '1')
-
-/**
- * Chroma (U/V) are subsampled horizontaly
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | ... |
- * | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | ...
- * | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | ...
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  U0/1      V0/1      U2/3      V2/3      ...
- *  U6/7      V6/7      U8/9      V8/9      ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_NV16 VIDEO_FOURCC('N', 'V', '1', '6')
-
-/**
- * Chroma (U/V) are subsampled horizontaly
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | ...
- * | ... |
- * | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | ...
- * | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | ...
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  V0/1      U0/1      V2/3      U2/3      ...
- *  V6/7      U6/7      V8/9      U8/9      ...
- *  ...
- * @endcode
- */
-
-#define VIDEO_PIX_FMT_NV61 VIDEO_FOURCC('N', 'V', '6', '1')
-
-/**
- * Chroma (U/V) are not subsampled
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | ... |
- * | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv |
- * | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv | Uuuuuuuu   Vvvvvvvv |
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  U0        V0        U1        V1        U2        V2        U3        V3        ...
- *  U6        V6        U7        V7        U8        V8        U9        V9        ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_NV24 VIDEO_FOURCC('N', 'V', '2', '4')
-
-/**
- * Chroma (U/V) are not subsampled
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | ... |
- * | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu |
- * | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu | Vvvvvvvv   Uuuuuuuu |
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  V0        U0        V1        U1        V2        U2        V3        U3        ...
- *  V6        U6        V7        U7        V8        U8        V9        U9        ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_NV42 VIDEO_FOURCC('N', 'V', '4', '2')
-
-/**
- * @brief Repeat a macro for every YUV Full-Planar format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_YUV_FULL_PLANAR(X, ...)					\
-	X(VIDEO_PIX_FMT_YUV420, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_YVU420, __VA_ARGS__)
-
-/**
- * Chroma (U/V) are subsampled horizontaly and vertically
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | ... |
- * | Uuuuuuuu | Uuuuuuuu |
- * | ... |
- * | Vvvvvvvv | Vvvvvvvv |
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  U0/1/6/7      U2/3/8/9      ...
- *  ...
- *
- *  V0/1/6/7      V2/3/8/9      ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_YUV420 VIDEO_FOURCC('Y', 'U', '1', '2')
-
-/**
- * Chroma (U/V) are subsampled horizontaly and vertically
- *
- * @code{.unparsed}
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy | Yyyyyyyy |
- * | ... |
- * | Vvvvvvvv | Vvvvvvvv |
- * | ... |
- * | Uuuuuuuu | Uuuuuuuu |
- * | ... |
- * @endcode
- *
- * Below diagram show how luma and chroma relate to each others
- *
- * @code{.unparsed}
- *  Y0        Y1        Y2        Y3        ...
- *  Y6        Y7        Y8        Y9        ...
- *  ...
- *
- *  V0/1/6/7      V2/3/8/9      ...
- *  ...
- *
- *  U0/1/6/7      U2/3/8/9      ...
- *  ...
- * @endcode
- */
-#define VIDEO_PIX_FMT_YVU420 VIDEO_FOURCC('Y', 'V', '1', '2')
-
-/**
- * @}
- */
-
-/**
- * @name Compressed formats
- * @{
- */
-
-/**
- * @brief Repeat a macro for every compressed format, passed as first parameter
- *
- * @param X macro to replicate
- * @param ... extra parameters
- */
-#define VIDEO_FOREACH_COMPRESSED(X, ...)					\
-	X(VIDEO_PIX_FMT_JPEG, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_H264, __VA_ARGS__)					\
-	X(VIDEO_PIX_FMT_H264_NO_SC, __VA_ARGS__)				\
-	X(VIDEO_PIX_FMT_PNG, __VA_ARGS__)
-
-/**
- * Both JPEG (single frame) and Motion-JPEG (MJPEG, multiple JPEG frames concatenated)
- */
-#define VIDEO_PIX_FMT_JPEG VIDEO_FOURCC('J', 'P', 'E', 'G')
-
-/**
- * H264 with start code
- */
-#define VIDEO_PIX_FMT_H264 VIDEO_FOURCC('H', '2', '6', '4')
-
-/**
- * H264 without start code
- */
-#define VIDEO_PIX_FMT_H264_NO_SC VIDEO_FOURCC('A', 'V', 'C', '1')
-
-/**
- * PNG
- */
-#define VIDEO_PIX_FMT_PNG VIDEO_FOURCC('P', 'N', 'G', ' ')
-
-/**
- * @}
- */
-
-/** @cond INTERNAL_HIDDEN */
-#define _VIDEO_FMT_OR_EQ(pixfmt_a, pixfmt_b) || ((pixfmt_a) == (pixfmt_b))
-/** @endcond */
-
-/**
- * @brief Test if a fourcc is a grayscale format
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_GRAYSCALE(pixfmt)						\
-	(0 VIDEO_FOREACH_GRAYSCALE(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is a Bayer format
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_BAYER(pixfmt)						\
-	(0 VIDEO_FOREACH_BAYER(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is an RGB format (bayer excluded)
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_RGB(pixfmt)						\
-	(0 VIDEO_FOREACH_RGB(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is an YUV format
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_YUV(pixfmt)						\
-	(0 VIDEO_FOREACH_YUV(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is any of the MIPI-packed formats
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_MIPI_PACKED(pixfmt)					\
-	(0 VIDEO_FOREACH_BAYER_MIPI_PACKED(_VIDEO_FMT_OR_EQ, pixfmt)		\
-	   VIDEO_FOREACH_GRAYSCALE_MIPI_PACKED(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is any of the MIPI-packed formats
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_PADDED(pixfmt)						\
-	(0 VIDEO_FOREACH_BAYER_PADDED(_VIDEO_FMT_OR_EQ, pixfmt)			\
-	   VIDEO_FOREACH_GRAYSCALE_PADDED(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is any of the semi-planar formats
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_SEMI_PLANAR(pixfmt)					\
-	(0 VIDEO_FOREACH_YUV_SEMI_PLANAR(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Test if a fourcc is any of the full-planar formats
- *
- * @param pixfmt FourCC of the pixel format to test
- * @return Whether the format is known to match this category
- */
-#define VIDEO_FMT_IS_FULL_PLANAR(pixfmt)					\
-	(0 VIDEO_FOREACH_YUV_FULL_PLANAR(_VIDEO_FMT_OR_EQ, pixfmt))
-
-/**
- * @brief Get number of bits per pixel of a pixel format
- *
- * @param pixfmt FourCC pixel format value (@ref video_pixel_formats).
- *
- * @retval 0 if the format is unhandled or if it is variable number of bits
- * @retval >0 bit size of one pixel for this format
- */
-static inline unsigned int video_bits_per_pixel(uint32_t pixfmt)
-{
-	switch (pixfmt) {
-	case VIDEO_PIX_FMT_SBGGR8:
-	case VIDEO_PIX_FMT_SGBRG8:
-	case VIDEO_PIX_FMT_SGRBG8:
-	case VIDEO_PIX_FMT_SRGGB8:
-	case VIDEO_PIX_FMT_GREY:
-		return 8;
-	case VIDEO_PIX_FMT_SBGGR10P:
-	case VIDEO_PIX_FMT_SGBRG10P:
-	case VIDEO_PIX_FMT_SGRBG10P:
-	case VIDEO_PIX_FMT_SRGGB10P:
-	case VIDEO_PIX_FMT_Y10P:
-		return 10;
-	case VIDEO_PIX_FMT_SBGGR12P:
-	case VIDEO_PIX_FMT_SGBRG12P:
-	case VIDEO_PIX_FMT_SGRBG12P:
-	case VIDEO_PIX_FMT_SRGGB12P:
-	case VIDEO_PIX_FMT_Y12P:
-	case VIDEO_PIX_FMT_NV12:
-	case VIDEO_PIX_FMT_NV21:
-	case VIDEO_PIX_FMT_YUV420:
-	case VIDEO_PIX_FMT_YVU420:
-		return 12;
-	case VIDEO_PIX_FMT_SBGGR14P:
-	case VIDEO_PIX_FMT_SGBRG14P:
-	case VIDEO_PIX_FMT_SGRBG14P:
-	case VIDEO_PIX_FMT_SRGGB14P:
-	case VIDEO_PIX_FMT_Y14P:
-		return 14;
-	case VIDEO_PIX_FMT_RGB565:
-	case VIDEO_PIX_FMT_YUYV:
-	case VIDEO_PIX_FMT_YVYU:
-	case VIDEO_PIX_FMT_UYVY:
-	case VIDEO_PIX_FMT_VYUY:
-	case VIDEO_PIX_FMT_SBGGR8P16:
-	case VIDEO_PIX_FMT_SGBRG8P16:
-	case VIDEO_PIX_FMT_SGRBG8P16:
-	case VIDEO_PIX_FMT_SRGGB8P16:
-	case VIDEO_PIX_FMT_SBGGR10:
-	case VIDEO_PIX_FMT_SGBRG10:
-	case VIDEO_PIX_FMT_SGRBG10:
-	case VIDEO_PIX_FMT_SRGGB10:
-	case VIDEO_PIX_FMT_SBGGR12:
-	case VIDEO_PIX_FMT_SGBRG12:
-	case VIDEO_PIX_FMT_SGRBG12:
-	case VIDEO_PIX_FMT_SRGGB12:
-	case VIDEO_PIX_FMT_SBGGR14:
-	case VIDEO_PIX_FMT_SGBRG14:
-	case VIDEO_PIX_FMT_SGRBG14:
-	case VIDEO_PIX_FMT_SRGGB14:
-	case VIDEO_PIX_FMT_SBGGR16:
-	case VIDEO_PIX_FMT_SGBRG16:
-	case VIDEO_PIX_FMT_SGRBG16:
-	case VIDEO_PIX_FMT_SRGGB16:
-	case VIDEO_PIX_FMT_Y8P16:
-	case VIDEO_PIX_FMT_Y10:
-	case VIDEO_PIX_FMT_Y12:
-	case VIDEO_PIX_FMT_Y14:
-	case VIDEO_PIX_FMT_Y16:
-	case VIDEO_PIX_FMT_NV16:
-	case VIDEO_PIX_FMT_NV61:
-		return 16;
-	case VIDEO_PIX_FMT_BGR24:
-	case VIDEO_PIX_FMT_RGB24:
-	case VIDEO_PIX_FMT_NV24:
-	case VIDEO_PIX_FMT_NV42:
-	case VIDEO_PIX_FMT_YUV24:
-		return 24;
-	case VIDEO_PIX_FMT_XYUV32:
-	case VIDEO_PIX_FMT_ARGB32:
-	case VIDEO_PIX_FMT_ABGR32:
-	case VIDEO_PIX_FMT_RGBA32:
-	case VIDEO_PIX_FMT_BGRA32:
-	case VIDEO_PIX_FMT_XRGB32:
-	case VIDEO_PIX_FMT_XBGR32:
-	case VIDEO_PIX_FMT_RGBX32:
-	case VIDEO_PIX_FMT_BGRX32:
-		return 32;
-	default:
-		/* Variable number of bits per pixel or unknown format */
-		return 0;
-	}
-}
-
-/**
- * @}
- */
 
 /**
  * @name MIPI CSI-2 Data Types
@@ -2525,4 +617,4 @@ static inline unsigned int video_bits_per_pixel(uint32_t pixfmt)
  * @}
  */
 
-#endif /* ZEPHYR_INCLUDE_VIDEO_H_ */
+#endif /* ZEPHYR_INCLUDE_DRIVERS_VIDEO_H */
