+++ samples/drivers/video/capture/src/main.c
@@ -9,10 +9,9 @@
 #include <zephyr/device.h>
 
 #include <zephyr/drivers/display.h>
-#include <zephyr/drivers/video.h>
-#include <zephyr/drivers/video-controls.h>
-
 #include <zephyr/logging/log.h>
+#include <zephyr/video/video.h>
+
 LOG_MODULE_REGISTER(main, CONFIG_LOG_DEFAULT_LEVEL);
 
 /*
