+++ samples/drivers/video/tcpserversink/src/main.c
@@ -7,17 +7,15 @@
  */
 
 #include <zephyr/device.h>
-#include <zephyr/drivers/video.h>
-#include <zephyr/drivers/video-controls.h>
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/net/socket.h>
-
-#include <zephyr/posix/netinet/in.h>
-#include <zephyr/posix/sys/socket.h>
 #include <zephyr/posix/arpa/inet.h>
-#include <zephyr/posix/unistd.h>
+#include <zephyr/posix/netinet/in.h>
 #include <zephyr/posix/poll.h>
+#include <zephyr/posix/sys/socket.h>
+#include <zephyr/posix/unistd.h>
+#include <zephyr/video/video.h>
 
 LOG_MODULE_REGISTER(main, CONFIG_LOG_DEFAULT_LEVEL);
