+++ subsys/Kconfig
@@ -60,6 +60,7 @@ source "subsys/usb/device/Kconfig"
 source "subsys/usb/device_next/Kconfig"
 source "subsys/usb/host/Kconfig"
 source "subsys/usb/usb_c/Kconfig"
+source "subsys/video/Kconfig"
 source "subsys/zbus/Kconfig"
 # zephyr-keep-sorted-stop
