+++ subsys/usb/common/uvc.c
@@ -5,7 +5,7 @@
 
 #include <stdint.h>
 
-#include <zephyr/drivers/video-controls.h>
+#include <zephyr/video/controls.h>
 #include <zephyr/drivers/video.h>
 #include <zephyr/sys/byteorder.h>
