+++ subsys/usb/device_next/class/usbd_uvc.c
@@ -17,7 +17,7 @@
 #include <zephyr/usb/usb_ch9.h>
 #include <zephyr/drivers/usb/udc.h>
 #include <zephyr/drivers/video.h>
-#include <zephyr/drivers/video-controls.h>
+#include <zephyr/video/controls.h>
 #include <zephyr/logging/log.h>
 
 #include <zephyr/sys/util.h>
