+++ subsys/video/Kconfig
@@ -1,0 +1,2 @@
+# SPDX-FileCopyrightText: The Zephyr Project Contributors
+# SPDX-License-Identifier: Apache-2.0
