+++ subsys/video/buffer.c
@@ -1,0 +1,16 @@
+/*
+ * SPDX-FileCopyrightText: The Zephyr Project Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <zephyr/drivers/video.h>
+#include <zephyr/logging/log.h>
+#include <zephyr/video/video.h>
+
+LOG_MODULE_REGISTER(video_buffer, CONFIG_VIDEO_LOG_LEVEL);
+
+int video_dequeue(const struct device *dev, struct video_buffer **vbuf,
+				       k_timeout_t timeout)
+{
+	return video_driver_dequeue(dev, vbuf, timeout);
+}
