+++ tests/drivers/video/test_pattern/src/main.c
@@ -9,11 +9,10 @@
 
 #include <zephyr/device.h>
 #include <zephyr/drivers/display.h>
-#include <zephyr/drivers/video-controls.h>
-#include <zephyr/drivers/video.h>
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/byteorder.h>
+#include <zephyr/video/video.h>
 #include <zephyr/ztest.h>
 
 LOG_MODULE_REGISTER(test_pattern, LOG_LEVEL_INF);
