+++ tests/lib/cpp/cxx/src/main.cpp
@@ -73,7 +73,6 @@
 #include <zephyr/drivers/uart.h>
 #include <zephyr/usb/usb_device.h>
 #include <zephyr/usb/class/usb_hid.h>
-#include <zephyr/drivers/video-controls.h>
 #include <zephyr/drivers/video.h>
 #include <zephyr/drivers/watchdog.h>
 
@@ -81,6 +80,11 @@
 #include <zephyr/rtio/rtio.h>
 #include <zephyr/sys/spsc_lockfree.h>
 #include <zephyr/sys/mpsc_lockfree.h>
+
+#include <zephyr/video/buffer.h>
+#include <zephyr/video/controls.h>
+#include <zephyr/video/formats.h>
+#include <zephyr/video/frmival.h>
 
 #include <zephyr/ztest.h>
